[2025-06-09] Lots of updates for build-utilities
DETAILS
-------

Added:
- NodeFiles methods (untested):
    - basename()
    - copyFile()
    - dirname()
    - getStats()
    - isDirectory()
    - isFile()
    - isSymLink()
    - mkdir()
    - readDir()
- Added cmd method to AbstractStage (in build scripts) - better error handling
- Types:
    - Json namespace with common schemas - TsConfig, PackageJson
    - KeysOptional & KeysRequired (with tests)
    - PartialExcept & RecursivePartialExcept (with tests)
    - RequiredPartially & RecursiveRequiredPartially (with tests)

Renamed & Moved:
- AbstractBuildStage.runStage() changed name to runSubStage()
- Meta type namespace was changed to Debug
- NodeFiles methods:
    - deleteFiles() → delete()
    - writeFiles() → write()
- NodeFiles types:
    - NodeFiles.Args.copyFileArgs → NodeFiles.Args.copyFile
    - NodeFiles.Args.readDirArgs → NodeFiles.Args.readDir
    - NodeFiles.Args.readFileArgs → NodeFiles.Args.readFile
    - NodeFiles.Args.writeFileArgs → NodeFiles.Args.writeFile
- The following type namespaces’ contents are now global:
    - Arrays, Functions, Node, StringLiterals

Removed (unused/redudant/etc):
- Removed distinguishArrays option from typeOf() and removed typeOf args completely (now empty)
- compileTypescript() method from AbstractStage (in build scripts)
- abstractArgs() static method from AbstractConfigurableClass
- abstractArgs() static method from AbstractBuildStage
- NodeFunctions class removed
- Functions class removed
- Types:
    - mergeArgs namespace
    - MergeObjects
    - Node namespace (now had no contents)
    - NodeConsole.CmdErrorHandler

Changed:
- Added constructor params optional param to AnyClass type
- Improved documentation
- Improved Classify type

Tests:
- Added tests to types:
    - Test (namespace)
    - Classify

+++ src/ts/classes/node/NodeFiles.ts
@@ -3,9 +3,6 @@
  * 
  * @packageDocumentation
  */
-/**
- * @package @maddimathon/utility-typescript@___CURRENT_VERSION___
- */
 /*!
  * @maddimathon/utility-typescript@___CURRENT_VERSION___
  * @license MIT
@@ -20,7 +17,7 @@ import NodePath from 'node:path';
 import { AbstractConfigurableClass } from '../abstracts/AbstractConfigurableClass.js';
 
 import {
-    mergeArgs,
+    arrayUnique,
 } from '../../functions/index.js';
 
 import { NodeConsole } from './NodeConsole.js';
@@ -28,6 +25,8 @@ import { NodeConsole } from './NodeConsole.js';
 
 /**
  * A configurable class for working with files and paths in node.
+ * 
+ * @since 0.2.0
  */
 export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
 
@@ -47,40 +46,37 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
     /* Args ===================================== */
 
     /**
+     * Default args for this stage.
+     * 
      * @category Args
      */
     public get ARGS_DEFAULT() {
 
+        const write = {
+            force: false,
+            rename: false,
+        };
+
         return {
-            argsRecursive: false,
-            root: './',
-            writeFileArgs: {
-                force: false,
-                rename: false,
+            argsRecursive: true,
+
+            copyFile: {
+                force: true,
+                rename: true,
+                recursive: false,
             },
+
+            root: './',
+
+            readDir: {
+                recursive: false,
+            },
+
+            readFile: {},
+
+            write,
+
         } as const satisfies NodeFiles.Args;
-    }
-
-    /**
-     * Build a complete args object.
-     * 
-     * @category Args
-     */
-    public buildArgs( args?: Partial<NodeFiles.Args> ): NodeFiles.Args {
-
-        const mergedDefault = AbstractConfigurableClass.abstractArgs(
-            this.ARGS_DEFAULT
-        ) as NodeFiles.Args;
-
-        // using this.mergeArgs here can cause issues because this method is 
-        // sometimes called from the prototype
-        const merged: NodeFiles.Args = mergeArgs(
-            mergedDefault,
-            args ?? {},
-            this.ARGS_DEFAULT.argsRecursive
-        );
-
-        return merged;
     }
 
 
@@ -97,6 +93,33 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
         super( args );
 
         this.nc = utils.nc ?? new NodeConsole( this.ARGS_DEFAULT as unknown as NodeConsole.Args );
+
+        const propNames = arrayUnique(
+            Object.getOwnPropertyNames( NodeFiles.prototype )
+                .concat( Object.getOwnPropertyNames( this.constructor.prototype ) )
+        ) as ( keyof NodeFiles | "constructor" )[];
+
+        for ( const _name of propNames ) {
+
+            // continues on match
+            switch ( _name ) {
+
+                case 'args':
+                case 'ARGS_DEFAULT':
+                case 'buildArgs':
+                case 'constructor':
+                case 'nc':
+                    continue;
+            }
+
+            // continues
+            if ( typeof this[ _name ] !== 'function' ) {
+                continue;
+            }
+
+            // @ts-expect-error
+            this[ _name ] = this[ _name ].bind( this );
+        }
     }
 
 
@@ -108,30 +131,102 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
     /* Files ===================================== */
 
     /**
+     * Copies a file to another path.
+     * 
+     * @category Filers
+     * 
+     * @experimental
+     * 
+     * @param source       Location to write file.
+     * @param destination  Location to copy the source path to.
+     * @param args         Optional configuration.
+     * 
+     * @return  Path to file if written, or false on failure.
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public copyFile(
+        source: string,
+        destination: string,
+        args: Partial<NodeFiles.CopyFileArgs> = {},
+    ): string | false {
+
+        source = this.pathResolve( source );
+        destination = this.pathResolve( destination );
+
+        args = this.mergeArgs( this.args.copyFile, args, false );
+
+        // returns if we aren't forcing or renaming
+        if ( this.exists( destination ) ) {
+
+            // returns
+            if ( !args.force && !args.rename ) {
+                return false;
+            }
+
+            if ( args.force ) {
+                this.delete( [ destination ] );
+            } else if ( args.rename ) {
+                destination = this.uniquePath( destination );
+            }
+        }
+
+        // returns
+        if ( !args.recursive && this.isDirectory( source ) ) {
+            this.mkdir( destination, { ...args, recursive: true } );
+            return this.exists( destination ) && destination;
+        }
+
+        const destinationDir = this.dirname( destination );
+
+        if ( !this.exists( destinationDir ) ) {
+            this.mkdir( destinationDir, { ...args, recursive: true } );
+        }
+
+        NodeFS.cpSync( source, destination, {
+            ...args,
+
+            errorOnExist: false,
+            mode: undefined,
+            preserveTimestamps: true,
+            // verbatimSymlinks: true,
+        } );
+
+        return ( this.exists( destination ) || this.isSymLink( destination ) ) && destination;
+    }
+
+    /**
      * Deletes given files.
      * 
      * @category Filers
      * 
      * @param paths         Paths to delete. Absolute or relative to root dir.
      * @param dryRun        If true, files that would be deleted are printed to the console and not deleted.
-     * @param logBaseLevel  Base depth for console messages (via NodeConsole).
+     * @param logLevel  Base depth for console messages (via NodeConsole).
+     * 
+     * @since ___PKG_VERSION___ — Renamed to delete from deleteFiles.
      */
-    public deleteFiles(
+    public delete(
         paths: string[],
+        logLevel: number = 0,
         dryRun: boolean = false,
-        logBaseLevel: number = 0,
     ): void {
 
         for ( const path of paths ) {
             // continues
             if ( !this.exists( path ) ) { continue; }
 
-            const stat = NodeFS.statSync( path );
+            const stat = this.getStats( path );
+
+            // continues
+            if ( !stat ) { continue; }
 
             if ( stat.isDirectory() ) {
 
                 if ( dryRun ) {
-                    this.nc.timestampLog( 'deleting directory: ' + path, { depth: logBaseLevel } );
+                    this.nc.timestampLog( 'deleting directory: ' + this.pathRelative( path ).replace( ' ', '%20' ), { depth: logLevel, linesIn: 0, linesOut: 0, maxWidth: null } );
                 } else {
                     NodeFS.rmSync( path, { recursive: true, force: true } );
                 }
@@ -139,12 +234,126 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
             } else if ( stat.isFile() || stat.isSymbolicLink() ) {
 
                 if ( dryRun ) {
-                    this.nc.timestampLog( 'deleting file: ' + path, { depth: logBaseLevel } );
+                    this.nc.timestampLog( 'deleting file: ' + this.pathRelative( path ).replace( ' ', '%20' ), { depth: logLevel, linesIn: 0, linesOut: 0, maxWidth: null } );
                 } else {
                     NodeFS.unlinkSync( path );
                 }
             }
         }
+    }
+
+    /**
+     * Gets the path dirname via {@link node:fs.dirname}.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public dirname( path: string ) {
+        return NodePath.dirname( path );
+    }
+
+    /**
+     * Gets the NodeFS.Stats value for the given path.
+     * 
+     * @category Meta
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public getStats( path: string ) {
+        if ( !this.exists( path ) ) {
+            return undefined;
+        }
+        return NodeFS.statSync( path );
+    }
+
+    /**
+     * Checks if the given path is a directory.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public isDirectory( path: string ) {
+        return this.getStats( path )?.isDirectory() ?? false;
+    }
+
+    /**
+     * Checks if the given path is a file.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public isFile( path: string ) {
+        return this.getStats( path )?.isFile() ?? false;
+    }
+
+    /**
+     * Checks if the given path is a symbolic link.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public isSymLink( path: string ) {
+        return this.getStats( path )?.isSymbolicLink() ?? false;
+    }
+
+    /**
+     * Creates a directory.
+     * 
+     * @category Filers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public mkdir( path: string, args?: NodeFS.MakeDirectoryOptions & {
+        recursive: true;
+    } ) {
+        return NodeFS.mkdirSync( path, args );
+    }
+
+    /**
+     * Read the paths within a directory.
+     * 
+     * @category Filers
+     * 
+     * @param path  Directory to read.
+     * @param args  Optional configuration.
+     * 
+     * @return  Paths within the given directory.
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public readDir(
+        path: string,
+        args: Partial<NodeFiles.ReadDirArgs> = {},
+    ): string[] {
+
+        args = this.mergeArgs( this.args.readDir, args, false );
+
+        return NodeFS.readdirSync(
+            this.pathResolve( path ),
+            {
+                ...args,
+                withFileTypes: false,
+                encoding: 'utf-8',
+            }
+        ).filter( path => !path.match( /(^|\/)\._/g ) );
     }
 
     /**
@@ -161,6 +370,8 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
         path: string,
         args: Partial<NodeFiles.ReadFileArgs> = {},
     ): string {
+
+        args = this.mergeArgs( this.args.readFile, args, false );
 
         return NodeFS.readFileSync(
             this.pathResolve( path ),
@@ -181,8 +392,10 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
      * @param args     Optional configuration.
      * 
      * @return  Path to file if written, or false on failure.
+     * 
+     * @since ___PKG_VERSION___ — Renamed to write from writeFiles.
      */
-    public writeFile(
+    public write(
         path: string,
         content: string | string[],
         args: Partial<NodeFiles.WriteFileArgs> = {},
@@ -195,7 +408,7 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
             : content;
 
         args = this.mergeArgs(
-            this.args.writeFileArgs,
+            this.args.write,
             args,
             false
         );
@@ -213,7 +426,7 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
             }
         }
 
-        NodeFS.mkdirSync( NodePath.dirname( path ), { recursive: true } );
+        this.mkdir( this.dirname( path ), { ...args, recursive: true } );
 
         NodeFS.writeFileSync( path, content, args );
 
@@ -241,7 +454,7 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
         const isRelative = !path.match( /^\//g );
 
         const newPath = this.pathResolve(
-            NodePath.dirname( path ),
+            this.dirname( path ),
             newName + NodePath.extname( path ),
         );
 
@@ -249,7 +462,36 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
     }
 
     /**
+     * Gets the basename of the given path.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
+     */
+    public basename(
+        path: string,
+        suffix?: string | false,
+    ): string {
+
+        if ( suffix === false ) {
+            suffix = undefined;
+        } else if ( !suffix ) {
+            suffix = NodePath.extname( path ) || undefined;
+        }
+
+        return NodePath.basename( path, suffix );
+    }
+
+    /**
      * Checks whether a file, directory, or link exists at the given path.
+     * 
+     * @category Path-makers
+     * 
+     * @since ___PKG_VERSION___
+     * 
+     * @experimental
      */
     public exists( path: string ): boolean {
         return NodeFS.existsSync( this.pathResolve( path ) );
@@ -264,7 +506,6 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
      */
     public pathRelative( path: string ): string {
         const relative = NodePath.relative( this.pathResolve(), path );
-
         return relative ? relative : '.';
     }
 
@@ -329,13 +570,40 @@ export class NodeFiles extends AbstractConfigurableClass<NodeFiles.Args> {
 
 /**
  * Used only for {@link NodeFiles}.
+ * 
+ * @since 0.2.0
  */
 export namespace NodeFiles {
 
     /**
      * Optional configuration for {@link NodeFiles}.
+     * 
+     * @since 0.2.0
      */
-    export type Args = AbstractConfigurableClass.Args & {
+    export interface Args extends AbstractConfigurableClass.Args {
+
+        argsRecursive: true;
+
+        /**
+         * Default configuration for {@link NodeFiles.copy}.
+         * 
+         * @since ___PKG_VERSION___ — Renamed to copyFile from copyFileArgs.
+         */
+        copyFile: CopyFileArgs;
+
+        /**
+         * Default configuration for {@link NodeFiles.readDir}.
+         * 
+         * @since ___PKG_VERSION___ — Renamed to readDir from readDirArgs.
+         */
+        readDir: ReadDirArgs;
+
+        /**
+         * Default configuration for {@link NodeFiles.readFile}.
+         * 
+         * @since ___PKG_VERSION___ — Renamed to readFile from readFileArgs.
+         */
+        readFile: ReadFileArgs;
 
         /**
          * Path to the root directory (relative to node's cwd).
@@ -345,23 +613,62 @@ export namespace NodeFiles {
         root: string;
 
         /**
-         * Default configuration for {@link NodeFiles.writeFile}.
+         * Default configuration for {@link NodeFiles.write}.
+         * 
+         * @since ___PKG_VERSION___ — Renamed to write from writeArgs.
          */
-        writeFileArgs: WriteFileArgs;
+        write: WriteFileArgs;
+    };
+
+    /**
+     * Optional configuration for {@link NodeFiles.copy}.
+     * 
+     * @see {@link https://nodejs.org/docs/latest-v22.x/api/fs.html#fscpsyncsrc-dest-options | node:fs.cpSync}
+     * 
+     * @since 0.2.0
+     */
+    export interface CopyFileArgs extends WriteFileArgs {
+
+        /**
+         * Function to filter copied files/directories.
+         *
+         * Return true to copy the item, false to ignore it.
+         *
+         * When ignoring a directory, all of its contents will be skipped as
+         * well.
+         */
+        filter?: ( src: string, dest: string ) => boolean;
+
+        /**
+         * Whether to copy directories recursively.
+         */
+        recursive: boolean;
+    };
+
+    /**
+     * Optional configuration for {@link NodeFiles.readDir}.
+     * 
+     * @since 0.2.0
+     */
+    export interface ReadDirArgs {
+        recursive: boolean;
     };
 
     /**
      * Optional configuration for {@link NodeFiles.readFile}.
+     * 
+     * @since 0.2.0
      */
-    export type ReadFileArgs = {
-        encoding: BufferEncoding;
+    export interface ReadFileArgs {
         flag?: string | undefined;
     };
 
     /**
-     * Optional configuration for {@link NodeFiles.writeFile}.
+     * Optional configuration for {@link NodeFiles.write}.
+     * 
+     * @since 0.2.0
      */
-    export type WriteFileArgs = Partial<ReadFileArgs> & {
+    export interface WriteFileArgs extends Partial<ReadFileArgs> {
 
         /** 
          * Overwrite file at destination if it exists.
